+++ CHANGELOG.md
@@ -17,6 +17,13 @@
 - ePortfolio endpoints (Thanks, [@Birdmaaan4](https://github.com/Birdmaaan4) and [@bennettscience](https://github.com/bennettscience))
 - Delete an Admin from an Account (Thanks, [@shaneros](https://github.com/shaneros))
 
+- Account Calendars
+  - List available account calendars
+  - List single account calendar
+  - Update a calendar's visibility
+  - Update many calendar's visibility
+  - List all account calendars
+
 ### General
 
 - Added support for Python 3.10